pyproject.toml
[project]
name = "tech-digest-bot"
version = "0.4.0"
description = "AI-powered tech news research bot with LangChain agent and Neo4j knowledge graph"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "Kunal Narang" }]
keywords = ["ai", "telegram", "bot", "tech-news", "langchain", "neo4j", "knowledge-graph", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "groq>=1.0.0",
    "ddgs>=9.0.0",
    "aiohttp>=3.9.0",
    # LangChain dependencies
    "langchain-core>=0.3.0",
    "langchain-neo4j>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-huggingface>=0.1.0",
    # Embeddings
    "sentence-transformers>=2.0.0",
    # Core dependencies
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "dependency-injector>=4.41.0",
    "jinja2>=3.1.0",
    "neo4j>=5.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
tech-digest-bot = "tech_digest_bot.bot.app:main"

[project.urls]
Repository = "https://github.com/yourusername/tech-digest-bot"
Documentation = "https://github.com/yourusername/tech-digest-bot/tree/main/docs"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tech_digest_bot"]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["ddgs", "telegram.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
